use protocols for runtime checks
